Modify for tigger to be SNS notification which includes details of the file. Place contents of zip at the same location as zip file was

+++ unzip.py
@@ -4,38 +4,39 @@ import urllib
 import zipfile
 import boto3
 import io
-
-
-print('Loading function')
+import json
+import os
 
 s3 = boto3.client('s3')
-bucket = 'my-bucket'
+bucket = ''
 
 def lambda_handler(event, context):
-    key = urllib.unquote_plus(event['Records'][0]['s3']['object']['key'].encode('utf8'))
-    try:
-        obj = s3.get_object(Bucket=bucket, Key=key)
-        putObjects = []
-        with io.BytesIO(obj["Body"].read()) as tf:
-            # rewind the file
-            tf.seek(0)
-
-            # Read the file as a zipfile and process the members
-            with zipfile.ZipFile(tf, mode='r') as zipf:
-                for file in zipf.infolist():
-                    fileName = file.filename
-                    putFile = s3.put_object(Bucket=bucket, Key=fileName, Body=zipf.read(file))
-                    putObjects.append(putFile)
-                    print(putFile)
-
-
-        # Delete zip file after unzip
-        if len(putObjects) > 0:
-            deletedObj = s3.delete_object(Bucket=bucket, Key=key)
-            print('deleted file:')
-            print(deletedObj)
-
-    except Exception as e:
-        print(e)
-        print('Error getting object {} from bucket {}. Make sure they exist and your bucket is in the same region as this function.'.format(key, bucket))
-        raise e
+    message = json.loads(event['Records'][0]['Sns']['Message'])
+    status = message['av-status']
+    key = urllib.unquote_plus(message['key'])
+    bucket = message['bucket']
+    path = os.path.dirname(key)
+    if status == 'CLEAN':
+        try:
+            obj = s3.get_object(Bucket=bucket, Key=key)
+            putObjects = []
+            with io.BytesIO(obj["Body"].read()) as tf:
+                # rewind the file
+                tf.seek(0)
+    
+                # Read the file as a zipfile and process the members
+                with zipfile.ZipFile(tf, mode='r') as zipf:
+                    for file in zipf.infolist():
+                        fileName = os.path.join(path, file.filename)
+                        putFile = s3.put_object(Bucket=bucket, Key=fileName, Body=zipf.read(file))
+                        putObjects.append(putFile)
+    
+    
+            # Delete zip file after unzip
+            if len(putObjects) > 0:
+                deletedObj = s3.delete_object(Bucket=bucket, Key=key)
+    
+        except Exception as e:
+            print(e)
+            print('Error getting object {} from bucket {}. Make sure they exist and your bucket is in the same region as this function.'.format(key, bucket))
+            raise e
